requirements.txt
Flask==2.3.3
PyMuPDF==1.23.8
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.2
chromadb==0.4.18
python-dotenv==1.0.0
gunicorn==21.2.0